Unify problem details handling, log parse errors
