+++ janus_collector/src/lib.rs
@@ -57,6 +57,7 @@ use derivative::Derivative;
 use http_api_problem::HttpApiProblem;
 use janus_core::{
     hpke::{self, associated_data_for_aggregate_share, HpkeApplicationInfo, HpkePrivateKey},
+    http::response_to_problem_details,
     retries::{http_request_exponential_backoff, retry_http_request},
     task::{url_ensure_trailing_slash, AuthenticationToken, DAP_AUTH_HEADER},
 };
@@ -69,7 +70,7 @@ use prio::{
 };
 use reqwest::{
     header::{HeaderValue, ToStrError, CONTENT_TYPE, LOCATION, RETRY_AFTER},
-    Response, StatusCode,
+    StatusCode,
 };
 use retry_after::FromHeaderValueError;
 use retry_after::RetryAfter;
@@ -78,7 +79,6 @@ use std::{
     time::{Duration as StdDuration, SystemTime},
 };
 use tokio::time::{sleep, Instant};
-use tracing::warn;
 use url::Url;
 
 /// Errors that may occur when performing collections.
@@ -355,13 +355,15 @@ where
                 if status == StatusCode::SEE_OTHER {
                     response
                 } else if status.is_client_error() || status.is_server_error() {
-                    return Err(response_to_error(response).await);
+                    return Err(Error::Http(response_to_problem_details(response).await));
                 } else {
                     return Err(Error::Http(HttpApiProblem::new(status)));
                 }
             }
             // Retryable error status code, but ran out of retries:
-            Err(Ok(response)) => return Err(response_to_error(response).await),
+            Err(Ok(response)) => {
+                return Err(Error::Http(response_to_problem_details(response).await))
+            }
             // Lower level errors, either unretryable or ran out of retries:
             Err(Err(error)) => return Err(Error::HttpClient(error)),
         };
@@ -416,13 +418,15 @@ where
                         return Ok(PollResult::NextAttempt(retry_after_opt));
                     }
                     _ if status.is_client_error() || status.is_server_error() => {
-                        return Err(response_to_error(response).await);
+                        return Err(Error::Http(response_to_problem_details(response).await));
                     }
                     _ => return Err(Error::Http(HttpApiProblem::new(status))),
                 }
             }
             // Retryable error status code, but ran out of retries:
-            Err(Ok(response)) => return Err(response_to_error(response).await),
+            Err(Ok(response)) => {
+                return Err(Error::Http(response_to_problem_details(response).await))
+            }
             // Lower level errors, either unretryable or ran out of retries:
             Err(Err(error)) => return Err(Error::HttpClient(error)),
         };
@@ -547,25 +551,6 @@ where
             .await?;
         self.poll_until_complete(&job).await
     }
-}
-
-/// Turn a [`reqwest::Response`] into an [`Error`]. If applicable, a JSON problem details document
-/// is parsed from the request's body, otherwise the error is solely constructed from the
-/// response's status code.
-async fn response_to_error(response: Response) -> Error {
-    let status = response.status();
-    if let Some(content_type) = response.headers().get(CONTENT_TYPE) {
-        if content_type == "application/problem+json" {
-            match response.json::<HttpApiProblem>().await {
-                Ok(mut problem) => {
-                    problem.status = Some(status);
-                    return Error::Http(problem);
-                }
-                Err(error) => warn!(%error, "Failed to parse problem details"),
-            }
-        }
-    }
-    Error::Http(HttpApiProblem::new(status))
 }
 
 #[cfg(feature = "test-util")]
