+++ janus_core/Cargo.toml
@@ -31,12 +31,13 @@ derivative = "2.2.0"
 futures = "0.3.24"
 hex = "0.4"
 hpke-dispatch = "0.4.0"
+http-api-problem = "0.55.0"
 janus_messages = { version = "0.1", path = "../janus_messages" }
 kube = { version = "0.65", optional = true, default-features = false, features = ["client", "rustls-tls"] }
 k8s-openapi = { version = "0.13.1", optional = true, features = ["v1_20"] }  # keep this version in sync with what is referenced by the indirect dependency via `kube`
 prio = { version = "0.10.0", features = ["multithreaded"] }
 rand = "0.8"
-reqwest = { version = "0.11.12", default-features = false, features = ["rustls-tls"] }
+reqwest = { version = "0.11.12", default-features = false, features = ["rustls-tls", "json"] }
 ring = "0.16.20"
 serde = { version = "1.0.145", features = ["derive"] }
 thiserror = "1.0"
