+++ janus_core/src/lib.rs
@@ -2,6 +2,7 @@ use std::future::Future;
 use tokio::task::JoinHandle;
 
 pub mod hpke;
+pub mod http;
 pub mod report_id;
 pub mod retries;
 pub mod task;
