+++ janus_server/src/aggregator/aggregation_job_driver.rs
@@ -1281,6 +1281,7 @@ mod tests {
         )]));
         let mocked_aggregate_failure = mock("POST", "/aggregate")
             .with_status(500)
+            .with_header("Content-Type", "application/problem+json")
             .with_body("{\"type\": \"urn:ietf:params:ppm:dap:error:unauthorizedRequest\"}")
             .create();
         let mocked_aggregate_success = mock("POST", "/aggregate")
@@ -1499,6 +1500,7 @@ mod tests {
         )]));
         let mocked_aggregate_failure = mock("POST", "/aggregate")
             .with_status(500)
+            .with_header("Content-Type", "application/problem+json")
             .with_body("{\"type\": \"urn:ietf:params:ppm:dap:error:unrecognizedTask\"}")
             .create();
         let mocked_aggregate_success = mock("POST", "/aggregate")
